update for use with v158 and v2xx
