--- YOUR_ADMIN/includes/functions/extra_functions/ezpages_metatags_admin_extra_functions.php
@@ -1,7 +1,0 @@
-<?php
-// -----
-// Part of the Meta Tags for EZ-Pages plugin -- prior to v4.0.0.
-//
-// This file is no longer used, serving **only** to overwrite the like-named file from a
-// previous version of the plugin, and can be safely removed from the store's file-system.
-//
