+++ zc_plugins/EzpagesMetatags/5.0.0/admin/includes/languages/english/extra_definitions/ezpages_metatags_extra_definitions.php
