+++ zc_plugins/EzpagesMetatags/5.0.0/manifest.php
@@ -1,0 +1,12 @@
+<?php
+return [
+    'pluginVersion' => '5.0.0',
+    'pluginName' => 'EZ Pages Metatags',
+    'pluginDescription' => 'Allows input of custom meta-tags for EZ Pages.',
+    'pluginAuthor' => 'lat9, pRoseLA',
+    'pluginId' => 0, // ID from Zen Cart forum
+    'zcVersions' => ['v158', 'v20x','v210',],
+    'changelog' => '',
+    'github_repo' => '',
+    'pluginGroups' => [],
+];
